requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-core
langchain-anthropic
langchain-openai
langchain-experimental
python-dotenv
langchain_ollama
reportlab
